Small file organization refactor
