+++ src/main/kotlin/dev/aasmart/models/PlayerData.kt
@@ -1,5 +1,6 @@
 package dev.aasmart.models
 
+import dev.aasmart.models.games.GameRole
 import kotlinx.serialization.Serializable
 
 @Serializable
